// ClientApp/src/app/shared/models/plant-master.ts
export interface PlantMasterModel {
    plantId: number;
    plantName: string;
    plantDesc: string;
    plantType: string;
    divisionId: number;
    regionId: number;
    active:string;
    divisionName:string;
    regionName:string;
    timezone:any;
    siteDirectorEmail:string;
    siteDirectorEmailHome:string;
}